+++ alfred-telemetry-platform/src/main/java/eu/xenit/alfred/telemetry/binder/AlfrescoStatusMetrics.java
@@ -6,14 +6,11 @@ import io.micrometer.core.instrument.binder.MeterBinder;
 import org.alfresco.repo.security.authentication.AuthenticationUtil;
 import org.alfresco.repo.transaction.RetryingTransactionHelper;
 import org.alfresco.service.cmr.admin.RepoAdminService;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import javax.annotation.Nonnull;
 
 public class AlfrescoStatusMetrics implements MeterBinder {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(AlfrescoStatusMetrics.class);
     private static final String STATUS_PREFIX = "alfresco.status";
 
     private RepoAdminService repoAdminService;
@@ -27,7 +24,6 @@ public class AlfrescoStatusMetrics implements MeterBinder {
 
     @Override
     public void bindTo(@Nonnull MeterRegistry meterRegistry) {
-        LOGGER.info("Registering Alfresco Status metrics");
         Gauge.builder(STATUS_PREFIX + ".readonly", repoAdminService, this::getReadOnly)
                 .description("Metric about Alfresco being in read-only mode")
                 .register(meterRegistry);
@@ -42,10 +38,6 @@ public class AlfrescoStatusMetrics implements MeterBinder {
                                 isReadOnly[0] = repoAdminService.getUsage().isReadOnly()),
                 true);
 
-        if (isReadOnly[0]) {
-            return 1d;
-        } else {
-            return 0d;
-        }
+       return (isReadOnly[0] ? 1d : 0d);
     }
 }